refactor nodes.go
